requirements.txt
pynput
Pillow
requests
python-dotenv
openai
pydub
simpleaudio

